1. List length

+++ 0x17-doubly_linked_lists/1-dlistint_len.c
@@ -10,7 +10,7 @@
 
 size_t dlistint_len(const dlistint_t *h)
 {
-	int node_count = 0;
+	size_t node_count = 0;
 
 	if (!h)
 	{
